+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Added the `CHOOSECOLS` dynamic-array function.
+
 ## [3.4.0] - 2026-08-10
 
 ### Added